refactor(routes): Remove lazy loading to be compatible witch itch.io servers
